LPD-75170 refactor: move all left files of ai-hub-rest-api to ai-hub-nexus-rest-api
