+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-api/bnd.bnd
@@ -1,0 +1,8 @@
+Bundle-Name: Liferay AI Hub Nexus REST API
+Bundle-SymbolicName: com.liferay.ai.hub.nexus.rest.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.ai.hub.nexus.rest.dto.v1_0,\
+	com.liferay.ai.hub.nexus.rest.manager.v1_0,\
+	com.liferay.ai.hub.nexus.rest.resource.v1_0,\
+	com.liferay.ai.hub.nexus.rest.resource.v1_0.util
