+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-api/src/main/resources/com/liferay/ai/hub/nexus/rest/resource/v1_0/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
