--- modules/dxp/apps/ai-hub/ai-hub-rest-api/src/main/java/com/liferay/ai/hub/rest/dto/v1_0/AgentDefinition.java
@@ -1,791 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.dto.v1_0;
-
-import com.fasterxml.jackson.annotation.JsonFilter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import com.liferay.petra.function.UnsafeSupplier;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
-import com.liferay.portal.vulcan.util.ObjectMapperUtil;
-
-import jakarta.annotation.Generated;
-
-import jakarta.validation.Valid;
-
-import jakarta.xml.bind.annotation.XmlRootElement;
-
-import java.io.Serializable;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.function.Supplier;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-@GraphQLName("AgentDefinition")
-@JsonFilter("Liferay.Vulcan")
-@XmlRootElement(name = "AgentDefinition")
-public class AgentDefinition implements Serializable {
-
-	public static AgentDefinition toDTO(String json) {
-		return ObjectMapperUtil.readValue(AgentDefinition.class, json);
-	}
-
-	public static AgentDefinition unsafeToDTO(String json) {
-		return ObjectMapperUtil.unsafeReadValue(AgentDefinition.class, json);
-	}
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	@Valid
-	public Map<String, Map<String, String>> getActions() {
-		if (_actionsSupplier != null) {
-			actions = _actionsSupplier.get();
-
-			_actionsSupplier = null;
-		}
-
-		return actions;
-	}
-
-	public void setActions(Map<String, Map<String, String>> actions) {
-		this.actions = actions;
-
-		_actionsSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setActions(
-		UnsafeSupplier<Map<String, Map<String, String>>, Exception>
-			actionsUnsafeSupplier) {
-
-		_actionsSupplier = () -> {
-			try {
-				return actionsUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	protected Map<String, Map<String, String>> actions;
-
-	@JsonIgnore
-	private Supplier<Map<String, Map<String, String>>> _actionsSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public Boolean getActive() {
-		if (_activeSupplier != null) {
-			active = _activeSupplier.get();
-
-			_activeSupplier = null;
-		}
-
-		return active;
-	}
-
-	public void setActive(Boolean active) {
-		this.active = active;
-
-		_activeSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setActive(
-		UnsafeSupplier<Boolean, Exception> activeUnsafeSupplier) {
-
-		_activeSupplier = () -> {
-			try {
-				return activeUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Boolean active;
-
-	@JsonIgnore
-	private Supplier<Boolean> _activeSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getDescription() {
-		if (_descriptionSupplier != null) {
-			description = _descriptionSupplier.get();
-
-			_descriptionSupplier = null;
-		}
-
-		return description;
-	}
-
-	public void setDescription(String description) {
-		this.description = description;
-
-		_descriptionSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setDescription(
-		UnsafeSupplier<String, Exception> descriptionUnsafeSupplier) {
-
-		_descriptionSupplier = () -> {
-			try {
-				return descriptionUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String description;
-
-	@JsonIgnore
-	private Supplier<String> _descriptionSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getExternalReferenceCode() {
-		if (_externalReferenceCodeSupplier != null) {
-			externalReferenceCode = _externalReferenceCodeSupplier.get();
-
-			_externalReferenceCodeSupplier = null;
-		}
-
-		return externalReferenceCode;
-	}
-
-	public void setExternalReferenceCode(String externalReferenceCode) {
-		this.externalReferenceCode = externalReferenceCode;
-
-		_externalReferenceCodeSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setExternalReferenceCode(
-		UnsafeSupplier<String, Exception> externalReferenceCodeUnsafeSupplier) {
-
-		_externalReferenceCodeSupplier = () -> {
-			try {
-				return externalReferenceCodeUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	protected String externalReferenceCode;
-
-	@JsonIgnore
-	private Supplier<String> _externalReferenceCodeSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public Long getId() {
-		if (_idSupplier != null) {
-			id = _idSupplier.get();
-
-			_idSupplier = null;
-		}
-
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-
-		_idSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setId(UnsafeSupplier<Long, Exception> idUnsafeSupplier) {
-		_idSupplier = () -> {
-			try {
-				return idUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	protected Long id;
-
-	@JsonIgnore
-	private Supplier<Long> _idSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	@Valid
-	public Variable[] getInputVariables() {
-		if (_inputVariablesSupplier != null) {
-			inputVariables = _inputVariablesSupplier.get();
-
-			_inputVariablesSupplier = null;
-		}
-
-		return inputVariables;
-	}
-
-	public void setInputVariables(Variable[] inputVariables) {
-		this.inputVariables = inputVariables;
-
-		_inputVariablesSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setInputVariables(
-		UnsafeSupplier<Variable[], Exception> inputVariablesUnsafeSupplier) {
-
-		_inputVariablesSupplier = () -> {
-			try {
-				return inputVariablesUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Variable[] inputVariables;
-
-	@JsonIgnore
-	private Supplier<Variable[]> _inputVariablesSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	@Valid
-	public Variable getOutputVariable() {
-		if (_outputVariableSupplier != null) {
-			outputVariable = _outputVariableSupplier.get();
-
-			_outputVariableSupplier = null;
-		}
-
-		return outputVariable;
-	}
-
-	public void setOutputVariable(Variable outputVariable) {
-		this.outputVariable = outputVariable;
-
-		_outputVariableSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setOutputVariable(
-		UnsafeSupplier<Variable, Exception> outputVariableUnsafeSupplier) {
-
-		_outputVariableSupplier = () -> {
-			try {
-				return outputVariableUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Variable outputVariable;
-
-	@JsonIgnore
-	private Supplier<Variable> _outputVariableSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	@Valid
-	public Status getStatus() {
-		if (_statusSupplier != null) {
-			status = _statusSupplier.get();
-
-			_statusSupplier = null;
-		}
-
-		return status;
-	}
-
-	public void setStatus(Status status) {
-		this.status = status;
-
-		_statusSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setStatus(
-		UnsafeSupplier<Status, Exception> statusUnsafeSupplier) {
-
-		_statusSupplier = () -> {
-			try {
-				return statusUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	protected Status status;
-
-	@JsonIgnore
-	private Supplier<Status> _statusSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getTitle() {
-		if (_titleSupplier != null) {
-			title = _titleSupplier.get();
-
-			_titleSupplier = null;
-		}
-
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
-
-		_titleSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setTitle(
-		UnsafeSupplier<String, Exception> titleUnsafeSupplier) {
-
-		_titleSupplier = () -> {
-			try {
-				return titleUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String title;
-
-	@JsonIgnore
-	private Supplier<String> _titleSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public Integer getVersion() {
-		if (_versionSupplier != null) {
-			version = _versionSupplier.get();
-
-			_versionSupplier = null;
-		}
-
-		return version;
-	}
-
-	public void setVersion(Integer version) {
-		this.version = version;
-
-		_versionSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setVersion(
-		UnsafeSupplier<Integer, Exception> versionUnsafeSupplier) {
-
-		_versionSupplier = () -> {
-			try {
-				return versionUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_ONLY)
-	protected Integer version;
-
-	@JsonIgnore
-	private Supplier<Integer> _versionSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getWorkflowDefinitionName() {
-		if (_workflowDefinitionNameSupplier != null) {
-			workflowDefinitionName = _workflowDefinitionNameSupplier.get();
-
-			_workflowDefinitionNameSupplier = null;
-		}
-
-		return workflowDefinitionName;
-	}
-
-	public void setWorkflowDefinitionName(String workflowDefinitionName) {
-		this.workflowDefinitionName = workflowDefinitionName;
-
-		_workflowDefinitionNameSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setWorkflowDefinitionName(
-		UnsafeSupplier<String, Exception>
-			workflowDefinitionNameUnsafeSupplier) {
-
-		_workflowDefinitionNameSupplier = () -> {
-			try {
-				return workflowDefinitionNameUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String workflowDefinitionName;
-
-	@JsonIgnore
-	private Supplier<String> _workflowDefinitionNameSupplier;
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof AgentDefinition)) {
-			return false;
-		}
-
-		AgentDefinition agentDefinition = (AgentDefinition)object;
-
-		return Objects.equals(toString(), agentDefinition.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		StringBundler sb = new StringBundler();
-
-		sb.append("{");
-
-		Map<String, Map<String, String>> actions = getActions();
-
-		if (actions != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"actions\": ");
-
-			sb.append(_toJSON(actions));
-		}
-
-		Boolean active = getActive();
-
-		if (active != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"active\": ");
-
-			sb.append(active);
-		}
-
-		String description = getDescription();
-
-		if (description != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"description\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(description));
-
-			sb.append("\"");
-		}
-
-		String externalReferenceCode = getExternalReferenceCode();
-
-		if (externalReferenceCode != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"externalReferenceCode\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(externalReferenceCode));
-
-			sb.append("\"");
-		}
-
-		Long id = getId();
-
-		if (id != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"id\": ");
-
-			sb.append(id);
-		}
-
-		Variable[] inputVariables = getInputVariables();
-
-		if (inputVariables != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"inputVariables\": ");
-
-			sb.append("[");
-
-			for (int i = 0; i < inputVariables.length; i++) {
-				sb.append(String.valueOf(inputVariables[i]));
-
-				if ((i + 1) < inputVariables.length) {
-					sb.append(", ");
-				}
-			}
-
-			sb.append("]");
-		}
-
-		Variable outputVariable = getOutputVariable();
-
-		if (outputVariable != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"outputVariable\": ");
-
-			sb.append(String.valueOf(outputVariable));
-		}
-
-		Status status = getStatus();
-
-		if (status != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"status\": ");
-
-			sb.append(String.valueOf(status));
-		}
-
-		String title = getTitle();
-
-		if (title != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"title\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(title));
-
-			sb.append("\"");
-		}
-
-		Integer version = getVersion();
-
-		if (version != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"version\": ");
-
-			sb.append(version);
-		}
-
-		String workflowDefinitionName = getWorkflowDefinitionName();
-
-		if (workflowDefinitionName != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"workflowDefinitionName\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(workflowDefinitionName));
-
-			sb.append("\"");
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	@io.swagger.v3.oas.annotations.media.Schema(
-		accessMode = io.swagger.v3.oas.annotations.media.Schema.AccessMode.READ_ONLY,
-		defaultValue = "com.liferay.ai.hub.rest.dto.v1_0.AgentDefinition",
-		name = "x-class-name"
-	)
-	public String xClassName;
-
-	private static String _escape(Object object) {
-		return StringUtil.replace(
-			String.valueOf(object), _JSON_ESCAPE_STRINGS[0],
-			_JSON_ESCAPE_STRINGS[1]);
-	}
-
-	private static boolean _isArray(Object value) {
-		if (value == null) {
-			return false;
-		}
-
-		Class<?> clazz = value.getClass();
-
-		return clazz.isArray();
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(_escape(entry.getKey()));
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			if (_isArray(value)) {
-				sb.append("[");
-
-				Object[] valueArray = (Object[])value;
-
-				for (int i = 0; i < valueArray.length; i++) {
-					if (valueArray[i] instanceof Map) {
-						sb.append(_toJSON((Map<String, ?>)valueArray[i]));
-					}
-					else if (valueArray[i] instanceof String) {
-						sb.append("\"");
-						sb.append(valueArray[i]);
-						sb.append("\"");
-					}
-					else {
-						sb.append(valueArray[i]);
-					}
-
-					if ((i + 1) < valueArray.length) {
-						sb.append(", ");
-					}
-				}
-
-				sb.append("]");
-			}
-			else if (value instanceof Map) {
-				sb.append(_toJSON((Map<String, ?>)value));
-			}
-			else if (value instanceof String) {
-				sb.append("\"");
-				sb.append(_escape(value));
-				sb.append("\"");
-			}
-			else {
-				sb.append(value);
-			}
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	private static final String[][] _JSON_ESCAPE_STRINGS = {
-		{"\\", "\"", "\b", "\f", "\n", "\r", "\t"},
-		{"\\\\", "\\\"", "\\b", "\\f", "\\n", "\\r", "\\t"}
-	};
-
-	private Map<String, Serializable> _extendedProperties;
-
-}
