--- modules/dxp/apps/ai-hub/ai-hub-rest-api/src/main/java/com/liferay/ai/hub/rest/dto/v1_0/Status.java
@@ -1,341 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.dto.v1_0;
-
-import com.fasterxml.jackson.annotation.JsonFilter;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import com.liferay.petra.function.UnsafeSupplier;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLField;
-import com.liferay.portal.vulcan.graphql.annotation.GraphQLName;
-import com.liferay.portal.vulcan.util.ObjectMapperUtil;
-
-import jakarta.annotation.Generated;
-
-import jakarta.xml.bind.annotation.XmlRootElement;
-
-import java.io.Serializable;
-
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.function.Supplier;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-@GraphQLName("Status")
-@JsonFilter("Liferay.Vulcan")
-@XmlRootElement(name = "Status")
-public class Status implements Serializable {
-
-	public static Status toDTO(String json) {
-		return ObjectMapperUtil.readValue(Status.class, json);
-	}
-
-	public static Status unsafeToDTO(String json) {
-		return ObjectMapperUtil.unsafeReadValue(Status.class, json);
-	}
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public Integer getCode() {
-		if (_codeSupplier != null) {
-			code = _codeSupplier.get();
-
-			_codeSupplier = null;
-		}
-
-		return code;
-	}
-
-	public void setCode(Integer code) {
-		this.code = code;
-
-		_codeSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setCode(UnsafeSupplier<Integer, Exception> codeUnsafeSupplier) {
-		_codeSupplier = () -> {
-			try {
-				return codeUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected Integer code;
-
-	@JsonIgnore
-	private Supplier<Integer> _codeSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getLabel() {
-		if (_labelSupplier != null) {
-			label = _labelSupplier.get();
-
-			_labelSupplier = null;
-		}
-
-		return label;
-	}
-
-	public void setLabel(String label) {
-		this.label = label;
-
-		_labelSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setLabel(
-		UnsafeSupplier<String, Exception> labelUnsafeSupplier) {
-
-		_labelSupplier = () -> {
-			try {
-				return labelUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String label;
-
-	@JsonIgnore
-	private Supplier<String> _labelSupplier;
-
-	@io.swagger.v3.oas.annotations.media.Schema
-	public String getLabel_i18n() {
-		if (_label_i18nSupplier != null) {
-			label_i18n = _label_i18nSupplier.get();
-
-			_label_i18nSupplier = null;
-		}
-
-		return label_i18n;
-	}
-
-	public void setLabel_i18n(String label_i18n) {
-		this.label_i18n = label_i18n;
-
-		_label_i18nSupplier = null;
-	}
-
-	@JsonIgnore
-	public void setLabel_i18n(
-		UnsafeSupplier<String, Exception> label_i18nUnsafeSupplier) {
-
-		_label_i18nSupplier = () -> {
-			try {
-				return label_i18nUnsafeSupplier.get();
-			}
-			catch (RuntimeException runtimeException) {
-				throw runtimeException;
-			}
-			catch (Exception exception) {
-				throw new RuntimeException(exception);
-			}
-		};
-	}
-
-	@GraphQLField
-	@JsonProperty(access = JsonProperty.Access.READ_WRITE)
-	protected String label_i18n;
-
-	@JsonIgnore
-	private Supplier<String> _label_i18nSupplier;
-
-	@Override
-	public boolean equals(Object object) {
-		if (this == object) {
-			return true;
-		}
-
-		if (!(object instanceof Status)) {
-			return false;
-		}
-
-		Status status = (Status)object;
-
-		return Objects.equals(toString(), status.toString());
-	}
-
-	@Override
-	public int hashCode() {
-		String string = toString();
-
-		return string.hashCode();
-	}
-
-	public String toString() {
-		StringBundler sb = new StringBundler();
-
-		sb.append("{");
-
-		Integer code = getCode();
-
-		if (code != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"code\": ");
-
-			sb.append(code);
-		}
-
-		String label = getLabel();
-
-		if (label != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"label\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(label));
-
-			sb.append("\"");
-		}
-
-		String label_i18n = getLabel_i18n();
-
-		if (label_i18n != null) {
-			if (sb.length() > 1) {
-				sb.append(", ");
-			}
-
-			sb.append("\"label_i18n\": ");
-
-			sb.append("\"");
-
-			sb.append(_escape(label_i18n));
-
-			sb.append("\"");
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	@io.swagger.v3.oas.annotations.media.Schema(
-		accessMode = io.swagger.v3.oas.annotations.media.Schema.AccessMode.READ_ONLY,
-		defaultValue = "com.liferay.ai.hub.rest.dto.v1_0.Status",
-		name = "x-class-name"
-	)
-	public String xClassName;
-
-	private static String _escape(Object object) {
-		return StringUtil.replace(
-			String.valueOf(object), _JSON_ESCAPE_STRINGS[0],
-			_JSON_ESCAPE_STRINGS[1]);
-	}
-
-	private static boolean _isArray(Object value) {
-		if (value == null) {
-			return false;
-		}
-
-		Class<?> clazz = value.getClass();
-
-		return clazz.isArray();
-	}
-
-	private static String _toJSON(Map<String, ?> map) {
-		StringBuilder sb = new StringBuilder("{");
-
-		@SuppressWarnings("unchecked")
-		Set set = map.entrySet();
-
-		@SuppressWarnings("unchecked")
-		Iterator<Map.Entry<String, ?>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, ?> entry = iterator.next();
-
-			sb.append("\"");
-			sb.append(_escape(entry.getKey()));
-			sb.append("\": ");
-
-			Object value = entry.getValue();
-
-			if (_isArray(value)) {
-				sb.append("[");
-
-				Object[] valueArray = (Object[])value;
-
-				for (int i = 0; i < valueArray.length; i++) {
-					if (valueArray[i] instanceof Map) {
-						sb.append(_toJSON((Map<String, ?>)valueArray[i]));
-					}
-					else if (valueArray[i] instanceof String) {
-						sb.append("\"");
-						sb.append(valueArray[i]);
-						sb.append("\"");
-					}
-					else {
-						sb.append(valueArray[i]);
-					}
-
-					if ((i + 1) < valueArray.length) {
-						sb.append(", ");
-					}
-				}
-
-				sb.append("]");
-			}
-			else if (value instanceof Map) {
-				sb.append(_toJSON((Map<String, ?>)value));
-			}
-			else if (value instanceof String) {
-				sb.append("\"");
-				sb.append(_escape(value));
-				sb.append("\"");
-			}
-			else {
-				sb.append(value);
-			}
-
-			if (iterator.hasNext()) {
-				sb.append(", ");
-			}
-		}
-
-		sb.append("}");
-
-		return sb.toString();
-	}
-
-	private static final String[][] _JSON_ESCAPE_STRINGS = {
-		{"\\", "\"", "\b", "\f", "\n", "\r", "\t"},
-		{"\\\\", "\\\"", "\\b", "\\f", "\\n", "\\r", "\\t"}
-	};
-
-	private Map<String, Serializable> _extendedProperties;
-
-}
