--- modules/dxp/apps/ai-hub/ai-hub-rest-api/src/main/resources/com/liferay/ai/hub/rest/dto/v1_0/packageinfo
@@ -1,1 +1,0 @@
-version 2.0.0
